requirements.txt
flask
tensorflow-cpu==2.10.0
opencv-python-headless
numpy
requests
gunicorn
